+++ src/web/UrlRule.php
@@ -2,7 +2,7 @@
 
 namespace davidhirtz\yii2\cms\web;
 
-use davidhirtz\yii2\cms\models\CategoryCollection;
+use davidhirtz\yii2\cms\models\collections\CategoryCollection;
 use davidhirtz\yii2\skeleton\web\Request;
 use davidhirtz\yii2\skeleton\web\UrlManager;