+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.4 - under development
 
 - Enhanced `SetupController` to also allow setting up categories
+- Moved `CategoryCollection` to `davidhirtz\yii2\cms\models\collections`
 - Removed `withFolder()` from `AssetQuery::withFiles()` query stack
 
 ## 2.1.3 (Jan 4, 2024)